scripts/demos/demo_qdrant_image_search_mock.py
#!/usr/bin/env python3
"""
Mock Demo: Qdrant Image Similarity Search (works without Qdrant)
Shows how the API works and what responses look like
"""

import json
from pathlib import Path

print("=" * 70)
print("🎨 QDRANT IMAGE SIMILARITY SEARCH - DEMO (Mock Mode)")
print("=" * 70)

print("\n📋 API Endpoint: POST /api/images/search-similar")
print("   Content-Type: multipart/form-data")
print("\n   Parameters:")
print("     - image_file (required): Image file to search")
print("     - limit (optional): Max results (default: 10)")
print("     - score_threshold (optional): Min similarity 0-1")

print("\n" + "=" * 70)
print("📤 Example Request (JavaScript)")
print("=" * 70)

example_request = """
const formData = new FormData();
formData.append('image_file', imageFile);
formData.append('limit', '10');
formData.append('score_threshold', '0.7');

const response = await fetch('/api/images/search-similar', {
    method: 'POST',
    body: formData
});

const results = await response.json();
console.log('Found', results.total_found, 'similar images');
"""

print(example_request)

print("\n" + "=" * 70)
print("📥 Example Response")
print("=" * 70)

example_response = {
    "query_image": "A yellow duckling with orange feet swimming in a blue pond with lily pads, soft watercolor style",
    "results": [
        {
            "image_uri": "http://example.org/book/duck_watercolor.png",
            "score": 0.95,
            "metadata": {
                "description": "A yellow duck with orange feet in watercolor style, soft pastel colors",
                "type": "output"
            }
        },
        {
            "image_uri": "http://example.org/book/duck_cartoon.png",
            "score": 0.87,
            "metadata": {
                "description": "A cute yellow duckling with big orange webbed feet, cartoon style",
                "type": "output"
            }
        },
        {
            "image_uri": "http://example.org/book/duck_family.png",
            "score": 0.82,
            "metadata": {
                "description": "A family of yellow ducklings with orange feet by a pond",
                "type": "output"
            }
        },
        {
            "image_uri": "http://example.org/book/duck_pond.png",
            "score": 0.78,
            "metadata": {
                "description": "A yellow duckling with orange feet swimming in a blue pond with lily pads",
                "type": "input"
            }
        }
    ],
    "total_found": 4
}

print(json.dumps(example_response, indent=2))

print("\n" + "=" * 70)
print("🌐 Frontend Integration")
print("=" * 70)

print("""
1. Open the HTML demo:
   http://localhost:8000/static/frontend_image_search_example.html

2. Or use in your own code:
   See docs/qdrant_frontend_integration.md for examples

3. React example:
   import { useState } from 'react';
   
   function ImageSearch() {
       const [results, setResults] = useState(null);
       
       const handleSearch = async (file) => {
           const formData = new FormData();
           formData.append('image_file', file);
           formData.append('limit', '10');
           
           const response = await fetch('/api/images/search-similar', {
               method: 'POST',
               body: formData
           });
           const data = await response.json();
           setResults(data);
       };
       
       return (
           <input type="file" onChange={(e) => handleSearch(e.target.files[0])} />
       );
   }
""")

print("\n" + "=" * 70)
print("🚀 To Run Full Demo")
print("=" * 70)

print("""
1. Start Qdrant:
   docker run -d -p 6333:6333 -p 6334:6334 qdrant/qdrant:latest

2. Start API server:
   python main.py

3. Run the demo:
   python demo_qdrant_image_search.py

4. Or test via curl:
   curl -X POST http://localhost:8000/api/images/search-similar \\
     -F "image_file=@path/to/image.jpg" \\
     -F "limit=10" \\
     -F "score_threshold=0.7"
""")

print("\n" + "=" * 70)
print("✅ Files Created")
print("=" * 70)

files_created = [
    "main.py - Added /api/images/search-similar endpoint",
    "static/frontend_image_search_example.html - Full HTML demo",
    "static/js/image_search_example.js - JavaScript library",
    "docs/qdrant_frontend_integration.md - Complete documentation",
    "demo_qdrant_image_search.py - Python demo script"
]

for i, file in enumerate(files_created, 1):
    print(f"   {i}. {file}")

print("\n" + "=" * 70)
print("🎯 Quick Test (when Qdrant is running)")
print("=" * 70)

print("""
# Test with Python requests:
import requests

with open('test_image.jpg', 'rb') as f:
    files = {'image_file': f}
    data = {'limit': '5', 'score_threshold': '0.7'}
    response = requests.post(
        'http://localhost:8000/api/images/search-similar',
        files=files,
        data=data
    )
    print(response.json())
""")

print("\n✅ Demo information displayed!")
print("   Ready to use when Qdrant is running.\n")

