requirements.txt
# Core dependencies
rdflib>=6.3.2
SPARQLWrapper>=2.0.0
pydantic>=2.5.2
pydantic-settings>=2.3.4
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Agent and AI dependencies
langchain>=0.0.350
openai>=1.3.7
transformers>=4.35.2
torch>=2.1.1

# Knowledge Graph dependencies
networkx>=3.2.1
owlready2>=0.37
pyshacl>=0.20.0

# Testing and development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
numpy>=1.24.0
tqdm>=4.66.1

# Vector Database and Search
qdrant-client>=1.7.0
tavily-python>=0.2.6

# Backoff library
backoff==2.2.1

# HTTP Client
httpx>=0.24.1
aiohttp>=3.8.0

# Google Cloud AI Platform & Auth
google-cloud-aiplatform>=1.48.0
google-auth>=2.28.0
google-api-python-client>=2.114.0
google-auth-oauthlib>=1.2.0
google-cloud-storage>=2.16.0
google-cloud-bigquery>=3.13.0

# Brain Rot Project Dependencies
pytrends>=4.9.2
better-profanity>=0.7.0
nltk>=3.8.1

# Async file operations
aiofiles>=23.1.0

# System monitoring
psutil>=5.9.6

# Metrics collection
prometheus_client>=0.17.1

# ... existing code ... 
twilio>=8.0.0 

# Added for diary_viewer_demo
nest_asyncio>=1.5.6

# Added for FastAPI file uploads
python-multipart>=0.0.20
Pillow
